The client now compiles and sort-of-but-not-really runs.
3d engine:
Ported to OSG 3.2 (although the camera has mostly been disabled).

2d GUI:
And ported to CEGUI 0.8.4 from who knows what.

It's also hard-coded to connect to a server running locally.

Also uses the latest osgCal.

And a lot of other miscellaneous updates.

None of this is confirmed to be working - right now the client only
shows blackness, but at least the mouse clicks appear to be working.
So the next thing to investigate will be the camera.